Refactor reference.ts to be one class per file.
