+++ dev/src/collection-group.ts
@@ -20,7 +20,8 @@ import * as protos from '../protos/firestore_v1_proto_api';
 import {QueryPartition} from './query-partition';
 import {requestTag} from './util';
 import {logger} from './logger';
-import {Query, QueryOptions} from './reference';
+import {Query} from './reference/query';
+import {QueryOptions} from './reference/query-options';
 import {FieldPath} from './path';
 import {Firestore} from './index';
 import {validateInteger} from './validate';
