+++ dev/src/reference/query-util.ts
@@ -1,0 +1,330 @@
+/**
+ * Copyright 2024 Google LLC. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import * as firestore from '@google-cloud/firestore';
+import {GoogleError} from 'google-gax';
+import {Duplex, Transform} from 'stream';
+
+import {Serializer} from '../serializer';
+import {Timestamp} from '../timestamp';
+import {DocumentSnapshotBuilder, QueryDocumentSnapshot} from '../document';
+import {
+  Deferred,
+  getTotalTimeout,
+  isPermanentRpcError,
+  requestTag,
+  wrapError,
+} from '../util';
+import {DocumentChange} from '../document-change';
+import {ExplainMetrics} from '../query-profile';
+import {logger} from '../logger';
+import {VectorQuery} from './vector-query';
+import {Query} from './query';
+import Firestore from '../index';
+import {QueryOptions} from './query-options';
+import {QuerySnapshot} from './query-snapshot';
+import {VectorQuerySnapshot} from './vector-query-snapshot';
+import {LimitType} from './types';
+import {NOOP_MESSAGE} from './constants';
+
+export class QueryUtil<
+  AppModelType,
+  DbModelType extends firestore.DocumentData,
+  Template extends
+    | Query<AppModelType, DbModelType>
+    | VectorQuery<AppModelType, DbModelType>,
+> {
+  constructor(
+    /** @private */
+    readonly _firestore: Firestore,
+    /** @private */
+    readonly _queryOptions: QueryOptions<AppModelType, DbModelType>,
+    /** @private */
+    readonly _serializer: Serializer
+  ) {}
+
+  _get(
+    query: Template,
+    transactionIdOrReadTime?: Uint8Array | Timestamp,
+    retryWithCursor = true
+  ): Promise<
+    | QuerySnapshot<AppModelType, DbModelType>
+    | VectorQuerySnapshot<AppModelType, DbModelType>
+  > {
+    const docs: Array<QueryDocumentSnapshot<AppModelType, DbModelType>> = [];
+
+    // Capture the error stack to preserve stack tracing across async calls.
+    const stack = Error().stack!;
+
+    return new Promise((resolve, reject) => {
+      let readTime: Timestamp;
+
+      this._stream(query, transactionIdOrReadTime, retryWithCursor)
+        .on('error', err => {
+          reject(wrapError(err, stack));
+        })
+        .on('data', result => {
+          readTime = result.readTime;
+          if (result.document) {
+            docs.push(result.document);
+          }
+        })
+        .on('end', () => {
+          if (this._queryOptions.limitType === LimitType.Last) {
+            // The results for limitToLast queries need to be flipped since
+            // we reversed the ordering constraints before sending the query
+            // to the backend.
+            docs.reverse();
+          }
+
+          resolve(
+            query._createSnapshot(
+              readTime,
+              docs.length,
+              () => docs,
+              () => {
+                const changes: Array<
+                  DocumentChange<AppModelType, DbModelType>
+                > = [];
+                for (let i = 0; i < docs.length; ++i) {
+                  changes.push(new DocumentChange('added', docs[i], -1, i));
+                }
+                return changes;
+              }
+            )
+          );
+        });
+    });
+  }
+
+  // This method exists solely to enable unit tests to mock it.
+  _isPermanentRpcError(err: GoogleError, methodName: string): boolean {
+    return isPermanentRpcError(err, methodName);
+  }
+
+  _hasRetryTimedOut(methodName: string, startTime: number): boolean {
+    const totalTimeout = getTotalTimeout(methodName);
+    if (totalTimeout === 0) {
+      return false;
+    }
+
+    return Date.now() - startTime >= totalTimeout;
+  }
+
+  stream(query: Template): NodeJS.ReadableStream {
+    if (this._queryOptions.limitType === LimitType.Last) {
+      throw new Error(
+        'Query results for queries that include limitToLast() ' +
+          'constraints cannot be streamed. Use Query.get() instead.'
+      );
+    }
+
+    const responseStream = this._stream(query);
+    const transform = new Transform({
+      objectMode: true,
+      transform(chunk, encoding, callback) {
+        callback(undefined, chunk.document);
+      },
+    });
+
+    responseStream.pipe(transform);
+    responseStream.on('error', e => transform.destroy(e));
+    return transform;
+  }
+
+  _stream(
+    query: Template,
+    transactionIdOrReadTime?: Uint8Array | Timestamp,
+    retryWithCursor = true,
+    explainOptions?: firestore.ExplainOptions
+  ): NodeJS.ReadableStream {
+    const tag = requestTag();
+    const startTime = Date.now();
+    const isExplain = explainOptions !== undefined;
+
+    let lastReceivedDocument: QueryDocumentSnapshot<
+      AppModelType,
+      DbModelType
+    > | null = null;
+
+    let backendStream: Duplex;
+    const stream = new Transform({
+      objectMode: true,
+      transform: (proto, enc, callback) => {
+        if (proto === NOOP_MESSAGE) {
+          callback(undefined);
+          return;
+        }
+
+        const output: {
+          readTime?: Timestamp;
+          document?: QueryDocumentSnapshot<AppModelType, DbModelType>;
+          explainMetrics?: ExplainMetrics;
+        } = {};
+
+        if (proto.readTime) {
+          output.readTime = Timestamp.fromProto(proto.readTime);
+        }
+
+        if (proto.document) {
+          const document = this._firestore.snapshot_(
+            proto.document,
+            proto.readTime
+          );
+          const finalDoc = new DocumentSnapshotBuilder<
+            AppModelType,
+            DbModelType
+          >(document.ref.withConverter(this._queryOptions.converter));
+          // Recreate the QueryDocumentSnapshot with the DocumentReference
+          // containing the original converter.
+          finalDoc.fieldsProto = document._fieldsProto;
+          finalDoc.readTime = document.readTime;
+          finalDoc.createTime = document.createTime;
+          finalDoc.updateTime = document.updateTime;
+          lastReceivedDocument = finalDoc.build() as QueryDocumentSnapshot<
+            AppModelType,
+            DbModelType
+          >;
+          output.document = lastReceivedDocument;
+        }
+
+        if (proto.explainMetrics) {
+          output.explainMetrics = ExplainMetrics._fromProto(
+            proto.explainMetrics,
+            this._serializer
+          );
+        }
+
+        callback(undefined, output);
+
+        if (proto.done) {
+          logger('QueryUtil._stream', tag, 'Trigger Logical Termination.');
+          backendStream.unpipe(stream);
+          backendStream.resume();
+          backendStream.end();
+          stream.end();
+        }
+      },
+    });
+
+    this._firestore
+      .initializeIfNeeded(tag)
+      .then(async () => {
+        // `toProto()` might throw an exception. We rely on the behavior of an
+        // async function to convert this exception into the rejected Promise we
+        // catch below.
+        let request = query.toProto(transactionIdOrReadTime, explainOptions);
+
+        let streamActive: Deferred<boolean>;
+        do {
+          streamActive = new Deferred<boolean>();
+          const methodName = 'runQuery';
+          backendStream = await this._firestore.requestStream(
+            methodName,
+            /* bidirectional= */ false,
+            request,
+            tag
+          );
+          backendStream.on('error', err => {
+            backendStream.unpipe(stream);
+
+            // If a non-transactional query failed, attempt to restart.
+            // Transactional queries are retried via the transaction runner.
+            // Explain queries are not retried with a cursor. That would produce
+            // incorrect/partial profiling results.
+            if (
+              !isExplain &&
+              !transactionIdOrReadTime &&
+              !this._isPermanentRpcError(err, 'runQuery')
+            ) {
+              logger(
+                'QueryUtil._stream',
+                tag,
+                'Query failed with retryable stream error:',
+                err
+              );
+
+              // Enqueue a "no-op" write into the stream and wait for it to be
+              // read by the downstream consumer. This ensures that all enqueued
+              // results in the stream are consumed, which will give us an accurate
+              // value for `lastReceivedDocument`.
+              stream.write(NOOP_MESSAGE, () => {
+                if (this._hasRetryTimedOut(methodName, startTime)) {
+                  logger(
+                    'QueryUtil._stream',
+                    tag,
+                    'Query failed with retryable stream error but the total retry timeout has exceeded.'
+                  );
+                  stream.destroy(err);
+                  streamActive.resolve(/* active= */ false);
+                } else if (lastReceivedDocument && retryWithCursor) {
+                  logger(
+                    'Query._stream',
+                    tag,
+                    'Query failed with retryable stream error and progress was made receiving ' +
+                      'documents, so the stream is being retried.'
+                  );
+
+                  // Restart the query but use the last document we received as
+                  // the query cursor. Note that we do not use backoff here. The
+                  // call to `requestStream()` will backoff should the restart
+                  // fail before delivering any results.
+                  if (this._queryOptions.requireConsistency) {
+                    request = query
+                      .startAfter(lastReceivedDocument)
+                      .toProto(lastReceivedDocument.readTime);
+                  } else {
+                    request = query.startAfter(lastReceivedDocument).toProto();
+                  }
+
+                  // Set lastReceivedDocument to null before each retry attempt to ensure the retry makes progress
+                  lastReceivedDocument = null;
+
+                  streamActive.resolve(/* active= */ true);
+                } else {
+                  logger(
+                    'QueryUtil._stream',
+                    tag,
+                    `Query failed with retryable stream error however either retryWithCursor="${retryWithCursor}", or ` +
+                      'no progress was made receiving documents, so the stream is being closed.'
+                  );
+                  stream.destroy(err);
+                  streamActive.resolve(/* active= */ false);
+                }
+              });
+            } else {
+              logger(
+                'QueryUtil._stream',
+                tag,
+                'Query failed with stream error:',
+                err
+              );
+              stream.destroy(err);
+              streamActive.resolve(/* active= */ false);
+            }
+          });
+          backendStream.on('end', () => {
+            streamActive.resolve(/* active= */ false);
+          });
+          backendStream.resume();
+          backendStream.pipe(stream);
+        } while (await streamActive.promise);
+      })
+      .catch(e => stream.destroy(e));
+
+    return stream;
+  }
+}
