+++ dev/src/reference/vector-query-options.ts
@@ -1,0 +1,36 @@
+/**
+ * Copyright 2024 Google LLC. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export class VectorQueryOptions {
+  constructor(
+    readonly limit: number,
+    readonly distanceMeasure: 'EUCLIDEAN' | 'COSINE' | 'DOT_PRODUCT'
+  ) {}
+
+  isEqual(other: VectorQueryOptions): boolean {
+    if (this === other) {
+      return true;
+    }
+    if (!(other instanceof VectorQueryOptions)) {
+      return false;
+    }
+
+    return (
+      this.limit === other.limit &&
+      this.distanceMeasure === other.distanceMeasure
+    );
+  }
+}
